# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SoulMatch AI - AI-powered matrimony platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /journey/answers:
    post:
      operationId: submitAnswer
      tags: [journey]
      summary: Submit answer to a journey question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "201":
          description: Answer submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JourneyAnswer"
  /journey/answers/{questionId}:
    put:
      operationId: updateAnswer
      tags: [journey]
      summary: Update an existing answer
      parameters:
        - name: questionId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerUpdate"
      responses:
        "200":
          description: Answer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JourneyAnswer"
...
  /interests:
    get:
      operationId: getInterests
      tags: [interests]
      summary: Get sent and received interests
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [sent, received, mutual]
      responses:
        "200":
          description: Interests
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Interest"
    post:
      operationId: sendInterest
      tags: [interests]
      summary: Send interest to a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterestInput"
      responses:
        "201":
          description: Interest sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interest"
  /interests/{interestId}:
    patch:
      operationId: respondToInterest
      tags: [interests]
      summary: Accept or decline an interest
      parameters:
        - name: interestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InterestResponse"
      responses:
        "200":
          description: Response recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Interest"
    delete:
      operationId: withdrawInterest
      tags: [interests]
      summary: Withdraw a sent interest
      parameters:
        - name: interestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Withdrawn
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string

    # ── AUTH SCHEMAS ────────────────────────────────────────────────────────
...
    Photo:
      type: object
      required: [id, url, isPrimary]
      properties:
        id:
          type: integer
        url:
          type: string
        isPrimary:
          type: boolean
        publicId:
          type: ["string", "null"]
...
    PublicProfile:
      type: object
      required: [id, firstName, age, city]
      properties:
        id:
          type: integer
        firstName:
          type: string
        displayName:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        occupation:
          type: ["string", "null"]
        education:
          type: ["string", "null"]
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        religion:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        photos:
          type: array
          items:
            $ref: "#/components/schemas/Photo"
        verificationStatus:
          type: string
        isPremium:
          type: boolean
        compatibilityScore:
          type: ["integer", "null"]
        journeyProgress:
          type: integer
...
    JourneyAnswer:
      type: object
      required: [id, questionId, answer, createdAt]
      properties:
        id:
          type: integer
        questionId:
          type: integer
        answer:
          type: string
        createdAt:
          type: string
    AnswerInput:
      type: object
      required: [questionId, answer]
      properties:
        questionId:
          type: integer
        answer:
          type: string
    AnswerUpdate:
      type: object
      required: [answer]
      properties:
        answer:
          type: string
...
    Interest:
      type: object
      required: [id, fromUserId, toUserId, status, createdAt]
      properties:
        id:
          type: integer
        fromUserId:
          type: integer
        toUserId:
          type: integer
        status:
          type: string
          enum: [pending, accepted, declined, withdrawn]
        message:
          type: ["string", "null"]
        createdAt:
          type: string
        fromUser:
          $ref: "#/components/schemas/PublicProfile"
        toUser:
          $ref: "#/components/schemas/PublicProfile"
    InterestInput:
      type: object
      required: [toUserId]
      properties:
        toUserId:
          type: integer
        message:
          type: string
    InterestResponse:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum: [accept, decline]
...